Extract AppTheme to own file and standardize Color+Hex
- Move AppTheme struct from TimerViewModel to Theme/AppTheme.swift (both apps)
- Extract Color(hex:) from macOS TimerViewModel to Extensions/Color+Hex.swift
- Fix notification name casing: iOS "TimerStatePollReceived" → "timerStatePollReceived"
- Remove 12 dead constants reintroduced during extraction (stripped in afc4160)
- Add Theme, Extensions, DateUtils, KeychainHelper to macOS Xcode target
