+++ packages/ios-app/Timetrack/Timetrack/Models/Models.swift
@@ -7,7 +7,7 @@ enum AppConstants {
 
 extension Notification.Name {
     static let idleTimeoutUpdated = Notification.Name("IdleTimeoutUpdated")
-    static let timerStatePollReceived = Notification.Name("TimerStatePollReceived")
+    static let timerStatePollReceived = Notification.Name("timerStatePollReceived")
 }
 
 // MARK: - User Models
