+++ packages/mac-app/TimeTrack/ViewModels/TimerViewModel.swift
@@ -783,93 +783,3 @@ class TimerViewModel: ObservableObject {
         return storedValue > 0 ? storedValue : AppConstants.defaultIdleTimeoutSeconds
     }
 }
-
-// MARK: - Premium App Theme
-struct AppTheme {
-    // Core colors - Rich, premium dark palette
-    static let primary = Color(hex: "#F8FAFC") ?? .white
-    static let secondary = Color(hex: "#94A3B8") ?? .secondary
-    static let tertiary = Color(hex: "#64748B") ?? .gray
-
-    // Accent colors - Vibrant gradient-ready
-    static let accent = Color(hex: "#6366F1") ?? .blue  // Indigo
-    static let accentSecondary = Color(hex: "#8B5CF6") ?? .purple  // Violet
-
-    // Backgrounds - Original dark palette
-    static let background = Color(hex: "#161516") ?? .black
-    static let backgroundElevated = Color(hex: "#1A1A1A") ?? .black
-    static let cardBackground = Color(hex: "#1F1F1F") ?? .gray
-    static let cardBackgroundHover = Color(hex: "#2A2A2A") ?? .gray
-
-    // Semantic colors - Rich and bold
-    static let success = Color(hex: "#22C55E") ?? .green
-    static let error = Color(hex: "#F43F5E") ?? .red  // Rose for premium feel
-    static let warning = Color(hex: "#F59E0B") ?? .orange
-
-    // Earnings highlight - Green for live money display
-    static let earnings = Color(hex: "#22C55E") ?? .green
-
-    // Border colors
-    static let border = Color(hex: "#27272A") ?? .gray
-
-    // Gradients
-    static let accentGradient = LinearGradient(
-        colors: [Color(hex: "#6366F1") ?? .blue, Color(hex: "#8B5CF6") ?? .purple],
-        startPoint: .topLeading,
-        endPoint: .bottomTrailing
-    )
-
-    static let successGradient = LinearGradient(
-        colors: [Color(hex: "#22C55E") ?? .green, Color(hex: "#10B981") ?? .green],
-        startPoint: .topLeading,
-        endPoint: .bottomTrailing
-    )
-
-    static let errorGradient = LinearGradient(
-        colors: [Color(hex: "#F43F5E") ?? .red, Color(hex: "#E11D48") ?? .red],
-        startPoint: .topLeading,
-        endPoint: .bottomTrailing
-    )
-
-    // Spacing system
-    static let spacingXS: CGFloat = 4
-    static let spacingSM: CGFloat = 8
-    static let spacingMD: CGFloat = 12
-    static let spacingLG: CGFloat = 16
-    static let spacingXL: CGFloat = 24
-    static let spacing2XL: CGFloat = 32
-
-    // Border radius
-    static let radiusSM: CGFloat = 6
-    static let radiusMD: CGFloat = 10
-    static let radiusLG: CGFloat = 14
-    static let radiusXL: CGFloat = 20
-}
-
-// MARK: - Color Extension
-extension Color {
-    init?(hex: String) {
-        let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        var int: UInt64 = 0
-        Scanner(string: hex).scanHexInt64(&int)
-        let a, r, g, b: UInt64
-        switch hex.count {
-        case 3: // RGB (12-bit)
-            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-        case 6: // RGB (24-bit)
-            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-        case 8: // ARGB (32-bit)
-            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-        default:
-            return nil
-        }
-
-        self.init(
-            .sRGB,
-            red: Double(r) / 255,
-            green: Double(g) / 255,
-            blue:  Double(b) / 255,
-            opacity: Double(a) / 255
-        )
-    }
-}
